+++ src/lib/mod.rs
@@ -1,1 +1,2 @@
 pub mod mail;
+pub mod seeds;