+++ tests/roles/mod.rs
@@ -1,0 +1,11 @@
+//! # Role Management
+//!
+//! Everything that is related to roles :
+//!
+//! - capabilities
+//! - role CRUD
+//! - add capability to role
+//! - assign role to user
+//!
+
+mod fetch;